[website-release] Consolidate firebase functions
